Crate som_parser_core

Modules

combinators

Generic parser combinators.

Structs

And

Sequences two parsers, one after the other, collecting both results.

AndLeft

Sequences two parsers, one after the other, but discards the output of the second one.

AndRight

Sequences two parsers, one after the other, but discards the output of the first one.

Map

Maps a function over the output value of the parser.

Or

Tries to apply the first parser, if it fails, it tries to apply the second parser.

Traits

Parser

Defines a parser.