som_parser_core

Module combinators

Generic parser combinators.

Enums

Either

Represents a value of either type A (Left) or type B (Right).

Functions

all

Applies every parser in a slice, from left to right, and returns the output from all of them. If one parser fails, the whole sequence is considered failed.

alternative

Tries to apply the first parser, if it fails, it tries to apply the second parser.

any

Tries every parser in a slice, from left to right, and returns the output of the first succeeding one.

between

Parses something that is enclosed between two other things.

default

Tries to apply a parser, or fallback to its default value (making it an always-succeeding parser).

either

Same as either, but allows for different output types for the parsers.

fallback

Tries to apply a parser, or fallback to a constant value (making it an always-succeeding parser).

many

Applies a parser zero or more times.

map

Transforms the output value of a parser.

not

Runs the given parser, fails if it succeeded, and succeeds otherwise.

optional

Tries to apply a parser, but fails gracefully (with an Option output).

peek

Transforms a parser into a non-consuming one, allowing to parse ahead without consuming anything.

sep_by

Parses zero or more things, separated by an arbitrary delimiter.

sep_by1

Parses one or more things, separated by an arbitrary delimiter.

sequence

Sequences two parsers, one after the other, collecting both results.

some

Applies a parser one or more times.