som_parser_core

Trait Parser

pub trait Parser<T, I, GCTXT>: Sized {
    // Required method
    fn parse(&mut self, input: I, genctxt: GCTXT) -> Option<(T, I, GCTXT)>;

    // Provided methods
    fn and<U, P: Parser<U, I, GCTXT>>(self, parser: P) -> And<Self, P> { ... }
    fn or<P: Parser<T, I, GCTXT>>(self, parser: P) -> Or<Self, P> { ... }
    fn map<F: Fn(T) -> U, U>(self, func: F) -> Map<Self, F, T> { ... }
    fn and_left<P: Parser<U, I, GCTXT>, U>(
        self,
        parser: P,
    ) -> AndLeft<Self, P, U> { ... }
    fn and_right<P: Parser<U, I, GCTXT>, U>(
        self,
        parser: P,
    ) -> AndRight<Self, P, T> { ... }
}

Defines a parser.

It is basically a function that takes an input and returns a parsed result along with the rest of input (which can be parsed further).

Required Methods

fn parse(&mut self, input: I, genctxt: GCTXT) -> Option<(T, I, GCTXT)>

Applies the parser on some input.

It returns the parsed value and the rest of the unparsed input as Some, if successful.
Failing that, it returns None.

Provided Methods

fn and<U, P: Parser<U, I, GCTXT>>(self, parser: P) -> And<Self, P>

Sequences two parsers, one after the other, collecting both results.

fn or<P: Parser<T, I, GCTXT>>(self, parser: P) -> Or<Self, P>

Tries to apply the first parser, if it fails, it tries to apply the second parser.

fn map<F: Fn(T) -> U, U>(self, func: F) -> Map<Self, F, T>

Maps a function over the output value of the parser.

fn and_left<P: Parser<U, I, GCTXT>, U>(self, parser: P) -> AndLeft<Self, P, U>

Sequences two parsers, one after the other, but discards the output of the second one.

fn and_right<P: Parser<U, I, GCTXT>, U>(self, parser: P) -> AndRight<Self, P, T>

Sequences two parsers, one after the other, but discards the output of the first one.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.